% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2tps.R
\name{xml2tps}
\alias{xml2tps}
\title{XML2TPS}
\usage{
xml2tps(xmlfile, tpsfile, fname = "", LM = 0, image_directory = "")
}
\arguments{
\item{xmlfile}{Name of the xml file}

\item{tpsfile}{Name of the tps file it will return}

\item{fname}{Part of the images name it will erase}

\item{LM}{Number of landmarks you wish to work}

\item{image_directory}{Directory of the images archive you will work}
}
\value{
A tps file with all the data specified in the xml file
}
\description{
XML2TPS
}
